mild 'but' testing refactor and basic clap tests
